Experiment 2 : Reimplement MP REST Client AsyncInvocationInterceptor support
Quite straightforward, the only thing that keeps the implementation complex is,
that PreInvocationInterceptor doesn't know what kind of invocation is going to happen,
therefore it cannot be registered on top level, and the decision to register interceptor
needs to be postponed until we know we're about to call async.

We could use one more flag in request context to signal that, but still we'd still
need to act in MethodModel.