Experiment 4 : Test context propagation with RxJava
This is where we fail. RxJava executes all of its processing outside
main thread, therefore even the interceptor gets executed on async
thread.

In ideal case it would be best if the interceptor can get control before
passing control to RxInvokerProvider, so we can guarantee it still
happens on calling thread.

While it is true, that rx request might not even fire ever, this is only
chance for service provider to propagate context.